requirements.txt
streamlit
mysql-connector-python
akshare
pandas